send email when form is submitted

+++ database/migrations/2020_01_22_180551_create_contacts_table.php
@@ -17,7 +17,7 @@ class CreateContactsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('email', 100);
-            $table->integer('phone')->nullable();
+            $table->unsignedBigInteger('phone')->nullable();
             $table->text('message');
             $table->timestamps();
